Refactored the ui directory structure
