+++ app/imports/ui/components/dashboard/dashboard-footer.html
