+++ app/imports/ui/components/dashboard/user-menu.js
