+++ app/imports/ui/components/index.js
@@ -1,9 +1,9 @@
 import './useraccounts';
-import './dashboard-header.html';
-import './dashboard-footer.html';
+import './dashboard/dashboard-header.html';
+import './dashboard/dashboard-footer.html';
 import './login-header.html';
 import './standard-header.html';
-import './simple_card.html'
+import './dashboard/simple_card.html'
 import './useraccounts';
-import './user-menu.html';
-import './user-menu.js';
+import './dashboard/user-menu.html';
+import './dashboard/user-menu.js';
